Nearing a calculation of the basics.
I think I have at least the general structure done with enough data at
this point to at least give calculation a try. The only real issue that
I've run into at this point is that Tactics is a negative modifier. For
that, we're going to start out at -50% and work our way up from there
with tactics then being positive like everything else when it's applied.

The only there thing to report is that we may have far too much that
could be done with this. I can even foresee branching this project off
into a whole website with shareable links with macro and skill gain
guides.